requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.1
huggingface-hub==0.24.0
idna==3.7
ipykernel==6.23.1
ipython==8.14.0
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.3
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==24.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
Send2Trash==1.8.0
sentence-transformers==3.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.13.0
threadpoolctl==3.5.0
tinycss2==1.1.1
tokenizers==0.19.1
torch==2.3.1
tornado==6.2
tqdm==4.66.4
traitlets==5.4.0
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==3.0.3
widgetsnbextension==4.0.3
